New GatedEditionCreator for editions v2
Editions v2 natively support time-limited open editions, so GatedEditionCreator no longer need an external TimeCop registry.

Additionally, the salt mechanism has changed in editions v2 so instead of using the normal edition factory, GatedEditionCreator itself becomes a factory (it directly clones and initializes an edition implementation). This allows it to compute the salt for the edition using the address in the SignedAttestation, not the msg.sender or tx.origin (which could be a relayer and not the actual creator).